refactor(tui): keep tab metadata focused
